채점 서버

+++ noticeSet/notedelete.php
@@ -12,13 +12,28 @@ if(!$_SESSION['id'] && !$_SESSION['name'] && !$_SESSION['pass']){
 
 $num = $_GET['num'];
 
+if($num == 0){
+	echo("
+		 <script>
+			 window.alert('액세스하는데 실패하였습니다.')
+			 history.go(-1);
+		 </script>
+		 ");
+	 exit;
+}
+
 $DOCUMENT_ROOT = $_SERVER["DOCUMENT_ROOT"];  // 웹 서버의 root 경로
 
 $fp = @fopen("$DOCUMENT_ROOT/pru.txt", "rb");
 
 if(!$fp) {
-  echo "파일 엑세스를 실패하였습니다";
-  exit;
+	echo("
+     <script>
+       window.alert('액세스하는데 실패하였습니다.')
+       history.go(-1);
+     </script>
+     ");
+   exit;
 }
 
 $dbpass = "";
@@ -35,6 +50,17 @@ $conn = mysqli_connect($servername,$username,$dbpasswd,$dbname);
 
 $sql = "SELECT * FROM notedata WHERE num = {$num}";
 $result=mysqli_query($conn,$sql);
+
+if($result == false){
+	echo("
+		 <script>
+			 window.alert('요청에 실패하였습니다.')
+			 history.go(-1);
+		 </script>
+		 ");
+	 exit;
+}
+
 $row = mysqli_fetch_assoc($result);
 
 
@@ -51,8 +77,24 @@ echo("
 $sql = "DELETE FROM noterow WHERE notenum = {$num}";
 $result=mysqli_query($conn,$sql);
 
+if($result == false){  echo("
+     <script>
+       window.alert('요청에 실패하였습니다.')
+       history.go(-1);
+     </script>
+     ");
+   exit;}
+
 $sql = "DELETE FROM notedata WHERE num = {$num}";
 $result=mysqli_query($conn,$sql);
+
+if($result == false){  echo("
+     <script>
+       window.alert('요청에 실패하였습니다.')
+       history.go(-1);
+     </script>
+     ");
+   exit;}
 
 echo("
 		<script>
